introduce chess game controller and refactor to support it
