+++ index.html
@@ -23,7 +23,8 @@
             <h2>Whites's Pieces Taken</h2>
             <div id="black-pieces"></div>
         </div>
-        <chess-board/></chess-board>
+        <chess-board id="main-game"/>
+        </chess-board>
         <moves-list class="moves-list"></moves-list>
     </div>
