+++ src/models/pieces/bishop.mjs
