+++ src/models/pieces/bishop.test.js
