+++ src/models/pieces/king.test.js
