+++ src/models/pieces/knight.mjs
