+++ src/models/pieces/knight.test.js
