+++ src/models/pieces/piece.mjs
@@ -1,4 +1,4 @@
-import eventBus from "../event-bus.js";
+import eventBus from "../../event-bus.js";
 import Move from "../move.js";
 import Board from "../board.js";
 
@@ -76,14 +76,14 @@ class Piece {
   }
 
   onMove(move) {
-    eventBus.dispatchEvent("piece-move", {
+    this.game.eventBus.dispatchEvent("piece-move", {
       pieceId: move.initiatingPiece.id,
       from: [move.sourceRow, move.sourceFile],
       to: [move.row, move.file],
       notation: move.toString()
     });
     if (move.capturedPiece) {
-      eventBus.dispatchEvent("piece-capture", {
+      this.game.eventBus.dispatchEvent("piece-capture", {
         pieceId: move.capturedPiece.id,
       });
     }
