+++ src/models/pieces/queen.mjs
