+++ src/models/pieces/rook.mjs
