+++ src/models/pieces/rook.test.js
